fixing the db errors

+++ Assets/Scripts/Conversation/NPC.cs
@@ -19,7 +19,7 @@ public class NPC : MonoBehaviour
     private int likeThreshold;
     private int neutralThreshold;
     private relationshipStatus relStat;
-    private int relationshipValue;
+    private double relationshipValue;
 
     // audio related
     [HideInInspector]
@@ -100,7 +100,7 @@ public class NPC : MonoBehaviour
         reader.Read();
         int nextPromptID = reader.GetInt32(0);
         string responseAudio = reader.IsDBNull(1) ? "" : reader.GetString(1);
-        int relationshipEffect = reader.GetInt32(2);
+        double relationshipEffect = reader.GetDouble(2);
         reader.Close();
 
         // Play the voice line for the response
@@ -157,7 +157,7 @@ public class NPC : MonoBehaviour
         reader.Read();
         promptID = reader.GetInt32(0);
         audioFolder = reader.GetString(1);
-        relationshipValue = reader.GetInt32(2);
+        relationshipValue = reader.GetDouble(2);
         dislikeThreshold = reader.GetInt32(3);
         neutralThreshold = reader.GetInt32(4);
         likeThreshold = reader.GetInt32(5);
@@ -338,7 +338,7 @@ public class NPC : MonoBehaviour
         }
     }
 
-    private void UpdateRelationshipValue(int relationshipEffect)
+    private void UpdateRelationshipValue(double relationshipEffect)
     {
         relationshipValue += relationshipEffect;
